handleStripe folder with apple pay and card options. Implemented inside backend. Not yet featured inside the front end; Added confirm button before deleting account inside 'verify email' page; translated password strenths in register page

+++ app/screens/EmailNotVerified.tsx
@@ -44,14 +44,15 @@ const EmailNotVerified = () => {
 
     const deleteAccount = () => {
 
-        Alert.prompt((t('delete-account')), (t('enter-password-to-delete-account')), [
+        Alert.alert((t('delete-account')), '', [
             {
-                text: 'Cancel',
+                text: t('cancel'),
                 style: 'cancel',
             },
             {
-                text: 'Delete',
-                onPress: async (password) => {
+                text: t('delete'),
+                style: 'destructive',
+                onPress: async () => {
                     deleteAccountDatabase();
                 }
             }
